Fixed the issue with the counter/pointer
I had an issue that the pointer did count correctly but the new state did not transfer to the prev and next functions accordingly and another click on the buttons was needed despite being on the last item in the array.
I decided to take the if statements out of the function and add an useEffect hook in order to make the pointer count correctly the jokes. I added a few if statements to make sure that all eventualities are covered and gave pointer and joke.length as dependencies
I will do a few more tests but it seems that everything is working
I also added comments to make sure that my code is easily understandable